+++ save-backend/src/main/resources/application-dev.properties
@@ -6,8 +6,13 @@ backend.orchestratorUrl=http://localhost:5100
 backend.demoUrl=http://localhost:5421
 backend.fileStorage.location=${user.home}/.save-cloud/cnb/files
 backend.file-storage.location=
-backend.s3-storage.endpoint=http://localhost:9000
-backend.s3-storage.bucketName=cnb
-backend.s3-storage.credentials.accessKeyId=admin
-backend.s3-storage.credentials.secretAccessKey=12345678
+#backend.s3-storage.endpoint=http://localhost:9000
+backend.s3-storage.endpoint=https://obs.ap-southeast-1.myhuaweicloud.com
+#backend.s3-storage.bucketName=cnb
+backend.s3-storage.bucketName=save-obs
+backend.s3-storage.prefix=cnb/files
+#backend.s3-storage.credentials.accessKeyId=admin
+backend.s3-storage.credentials.accessKeyId=QC0EEHXWJDWJMBOX59TW
+#backend.s3-storage.credentials.secretAccessKey=12345678
+
